requirements/base.txt
fastapi==0.95.2
uvicorn==0.22.0
websockets==11.0.3
pydantic[dotenv]==1.10.8
cloudinary==1.33.0

neomodel==5.0.0

# GraphQL
# ------------------------------------------------------------------------------
# graphql-core<3
graphene==3.2.2
starlette-graphene3==0.6.0
graphene-file-upload==1.3.0

# JWT
# ------------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
